Initial IntAct parser
The data model is pretty complex but this adds an initial parser for
their data. I did have to add a PEG parser to more easily deal with the
quoting and escaping rules they have. Hopefully this doesn't make the
overall parsing too slow. Also this adds some basic tests to ensure that
it can parse the data, but does not yet have the required logic for
loading the results into the database.