Do not group ENA files
Seems some files are causing crashes. Not sure why or how but it looks
like there are bad bytes somewhere within the file. I can't track down
which file this is, so for now we will just process each individually.
This should help find the bugs, but will cause a lot of jobs (~200k),
but since nextflow will never submit for than 2k at once it shouldn't be
too bad.