Initial streaming lookup
The idea behind this is not to try and index the xml, but instead to
just parse it and write out the requested data as we parse. This
requires the input file to expand to be small enough to fit in memory
easily, but that shouldn't be a problem. I leave the old indexing code
around as it may still prove useful. Also this is not yet tested or part
of the pipeline since that will require some more work.

Also, along the way I tracked down a memory issue from the XML parsing.
It was still building a large XML tree even though I didn't want it to.
This now builds a large tree but with empty elements. This should be
much faster and use less memory.