Try to rewire pipeline to stream lookup pubs
The publication lookup is slow right now. This is likely because of some
file system slowness. This tries to fix that by doing less work on the
file system. We can basically skip indexing and instead chunk the
references to lookup and just stream over the XML files to find the
sequences. This should also increase the general pipeline speed.