Fix most tests
Since we are not using RNA types as SO terms the tests need to be
updated. Not 100% sure what to do with the 'other' term right now.
According to INSDC it is:

http://www.sequenceontology.org/browser/current_release/term/SO:0000110

which isn't that useful. I'm going to consider more. I might move it to
the top level ncRNA type instead.