Attempt to deal with quoting issues
It seems that python's CSV module will strip quotes off from around a
field. This is a problem for us if we have a field like:

"escaped(key):value"|another:"escaped:value"

As stripping the quotes produces something that cannot be parsed
correctly. If we prevent the quote stripping it works.