OUtline creating intact sequences
This outlines the logic needed to create sequence entries for intact
data. Basically we do something similar to GeneCards in that we use
their URS_taxid info to lookup the required information and then create
a sequence entry for it. This is a little different in that I do the
lookup in while parsing, which I do not like.

Also, this adds a new type of file and data to the entry objects,
interactions.