Try to speed up intact fetching
It takes roughly forever to parse intact data. This because the parser
for each id reference they have is way too slow. Instead of removing it
or not parsing things I can just grep the file ahead of time to pull out
lines that may be interesting. To select interactions we look for
rnacentral ids (which are always of the form rnacentral:URSNNN) in
various fields. By just grepping for rnacentral anywhere in the line we
should get the same results but much much faster.