Add logic to validate sequence counts
We should check that the numbers of new sequences that will be created
is reasonable. We do not want to flood the database with too many new
sequences and the like each release. This adds a check to see if the
counts, by database of new sequences is within a reasonable limit.
Currently it defaults to a 30% increase for each database. Hopefully
this helps us catch issues with ENA and sequence fragments easier.