Load translated SO terms into precomputed
We should load translated SO terms into the precomptued table. This will
allow us to start displaying them in places. This is the start of the
switch to SO terms. For now we can just use existing terms but as we go
we can start computing better terms and comparing them as needed.