Initial tree computation logic
This can compute a simple SO term tree that we can export. The idea is
to have some SO terms indexed in a hierarchy for the search and display
on the website. This provides the logic to compute the tree needed. It
will be used as part of the metadata step for indexing eventually.

This has some limits right now. First it assumes that all entries are
ncRNAs, and this will not be true when we have precursor molecules.
Secondly, it assumes there is always a single path. That may be a limit
of how indexing will be done but it may be an issue later. Either way
this is a reasonable first step at creating the logic to produce the
required trees.