Try to write out SO term tree
This is an attempt at writing out the SO term tree into the xml for the
search team. This has some other useful changes, moving the merging of
extra annotations. Maybe that is a way forward on improving the
efficiency of search export.

This tries to create the SO term tree to search. It uses names, not the
SO ids. May not be ideal. Also, this does not allow for aliases (lncRNA
or lnc_RNA), which way may want.