Try to index search metadata
I can index the search metadata with minimal changes to the interface, I
think. This attempts to do so. This is getting required since the export
is now taking silly amounts of time and memory (>30Gb). This should cut
down on memory. Hopefully something can be done to fix the time
issues.