Try to improve metadata caching
We don't need to store an SO term tree for each sequence, we can just
build a tree for each possible entry and then look it up as needed.
Basically we were computing and storing tons of duplicate data, which
this should eliminate most, if not all of. This should improve both
memory and run time of the search export pipeline.