Update to next R2DT data format
Soon there will be a more organized format for the data. This updates to
parse that new format.

We also determine the model_id while parsing to remove the need for
joins when inserting. This could allow for more easily inserting data
into the database as we do not require a temp loading table to do so.
This also allows for easier validation of the assumptions I have about
the data (model table is up-to-date).