Implement file based publishing
This adds a process to publish traveler SVG files to a directory. The
idea is that we have some stable setup for publishing the files that is
faster and easier than the database. A tree of files works well and is
easy to implement. This should create a reasonable tree for us.

Note I limit the max forks of publishing to 5 right now. This is random
but we need a limit to not cause too much file system strain. This
should still be fast enough for general usage.