First pass at generating CRW data
This can parse the CRW data into a form for import. It is not integrated
into the pipeline proper, and likely won't be. The reason is that we are
likely to only do this parsing as a one off task, and not something that
needs to happen again. If we ever have to redo it, it can be integrated
more throughly then.