Index all sequences with all parent terms
This indexes all entries with all assigned and inferable SO terms. The
goal behind this is that we want all searches to work, so you can search
for a subclass and find it or search for a parent of it and find it as
well. This might obscure some search results (things only assigned to a
parent term), but that is ok for us we think.