Add missing queries
Making some minor modifications to the queries that extract coordinates
from RNAcentral for FTP export for the CRS work. We basically need to
use this to filter out the Rfam only CRS as well as select the CRS
features that match in a sequence. As part of this I move the loading
code into files/crs, as well as remove the unneeded schema creation.

Also fix some typos in the pipeline.