Try to speed up qa sequence fetching
This is an attempt to speed up how long it takes to get the sequences to
run QA on. Currently, this is the slowest part of the pipeline. It
generally takes a while, but lately it has taken over a day to compute
the sequences. This is not acceptable and so this tries to fix it.

What this does is compute the urs ids to use outside the database, which
can be fast when they are sorted. We then use those ids to pull out the
sequences we want. We also avoid using seqkit at all and generate the
chunks and convert from JSON to fasta in a single step using split. I
can speed up the json2fasta.py script, with a simple rust version if
need be, but hopefully this is enough of an improvement.

This approach loses the shuffling, but this can probably be recreated by
using `--number r/${params.qa[name].chunk_size}` and changing the filter
to append (I think?). That isn't a perfect randomizing, but doing a
round robin shuffle should be close enough if it has to happen.

I may need to up the memory requirements and fiddle with the method of
importing urs to process, but this should cover a much faster method.