Start splitting apart large queries
One of the problems with search export and precompute is how slow it is
to extract data from the database. This begins work to fix that. I
create what I think are reasonable small queries that can be run and
used to rebuild the data needed for search export/precompute. These
queries are slightly different in that some things are logically
duplicated and I try to remove those parts.

This also assumes that there will be an numeric id column used to refer
to urs/taxid combinations. Not sure if this is a good idea, or needed.
Though it does make some things mildly easier.