Begin limiting container usage and mounting
One of the reason for poor performance of various steps is the fact that
the cluster can and will cause random hangs. This is particularly a
problem for search export since there are several processes, but only
one occurs a time so a single hang freezes everything forever.

One way to deal with this is to not use containers. This means we now
assume that psql is present in the path. This is ok on the cluster since
it is.

The other fix is to limit what is mounted. By limiting the mounts what
is mounted, we don't have to worry about other filesystems issues
causing the pipeline to hang.

These changes should fix, or at least those issues. The overall slowness
remains, but now it should not be due to external factors.