Start outlining gene loading
This is the process that I think will occur. It is very much a not
working outline right now but does cover more or less what I think
needs to happen. This covers the basic steps, getting the data,
splitting it by rna_type/chromosome/strand and then grouping things as
needed to create 'genes'.

I've copied the query from building bed/gff files for now. I suspect
that this will need more information than that as time goes, but for now
this should be plenty start with.