Initial attempt at grouping
As some point we will actually group sequences on the genome to produce
genes. This does start some grouping work but doesn't do anything
different in terms of logic for detecting if something is representative
or not. This just tries to group things, constantly expanding the size,
and then mark all the ones without QA issues as representative.