Outline idea of improving ENA parsing
ENA parsing in the pipeline isn't ideal right now. In the current
approach I either have too many (>100k) jobs, or have to merge then
split files. Neither of these are ideal. We can do a bit better by using
the fact that ENA puts the files into directories of reasonablish size. This
work outlines what will need to happen to use that order in the python
code. We will fetch some files directly and for other parts we will use
the directory structure to process one directory at a time. This is just
an outline and doesn't work yet.