Split R2DT data twice
WE should split twice. The reason being we need to ensure that we never
produce a directory with more than a few thousand files. Having too many
files stresses the file system out a lot. If we split once into small groups
then we have an issue with the split directory. If we split into large
groups then we have a problem with each R2DT output directory. Since it
produces many files per sequence having too many sequences in a chunk
will be a problem. By splitting twice we should limit the number of
files to something reasonable.