Rework output writing
We want to write both the rejected and grouped things. I think the best
way of doing this is by having more than one file, which means a large
change to the output work. This can now write more than one file in the
requested format. This will make it easier for me to see how things are
classified as well as load things into the database.