Begin work on hiding Rfam overlaps
This outlines the code needed to hide any Rfam hit overlapping another
sequence. This also features a large rewrite of how I think of some of
the data. As a small note I split the data file apart because it was
annoyingly large.

First, we need to add another type of location type, ignored. This is
used for things that we never really tried to fit into genes, eg lncRNA
right now. We want this to track things that aren't cluster yet so we
should store the locations so we can import that they were not
grouped.data. I also change the name of Locus to Cluster to reflect that
is a cluster of sequences, as well as change from 'representative' to
'highlighted', as this is less incorrect about our intentions. Probably
not final, but is at least closer I hope.

The rewrite now features a more complex logic. The idea is to first
reject clearly useless things (eg incomplete RNA sequences), then build
up all overlapping clusters, then process and split each cluster. We
just group things together if they overlap regardless of the RNA type.
Here we can do some processing of any cluster regardless of RNA type,
for example rejecting all clusters that overlap a known pseudogene. Then
we can split them based on RNA type and strand as needed. This allows us
to place misc/other that may be rRNA with an rRNA cluster and so forth.
From there we can analyze for Rfam overlaps and reject what needs to
rejected. Once that is done we than analyze each cluster with RNA type
specific code.

The idea behind these changes is that we need some generic steps prior
to RNA type specific steps, and that deciding which sequence a cluster
should go in is very RNA type and species specific. This provides some
simplish places to extend the logic to handle this.

The code isn't tested yet and likely doesn't run as it just my outline
of what will happen.