Large rewrite
The previous version had some design level issues that I couldn't solve.
This reworks the code to have a better, hopefully design. Some of these
issues were, ability to easily add generic reasons for not clustering
sequences (they are a pseudogene), ability to group things together
when they have different, but related SO terms, preventing the building
of overly large clusters, and ensuring that each location is only ever
in one cluster. Additionally, we decided to add a new state for
locations, ignored. This is meant to show which things we never really
tried to cluster. So the states a sequence can be in are: clustered
(sort of a gene), rejected (has an issue that means it should be
always hidden) and ignored (we never worked with it).

Additionally, each member of a cluster can be highlighted or not.
Highlighted means we thing the sequence is 'good' or worth showing to
users somehow. Ones which are not have some sort of issue and shouldn't
be shown unless requested.

The new design is basically to first exclude things if we believe them
to have issues. For example, we ignore all lncRNAs right now.

Then we group things that they overlap with if they have a compatible
RNA type. This is determined by selecting specific SO terms as going
searching the SO ontology for one. If one is found we use that for
grouping purposes. If none is found we just use the term itself. Every
overlapping location that has the same term (normalized or
not) are put together. We build up all clusters on a chromosome like
this.

We then process each cluster to first reject any sequences which only
come from Rfam but overlap something else. Then we do RNA type specific
logic. Currently anything which is a child of rRNA is processed and
everything else is ignored.

This does not solve all issues. Some are:

1. Issues with rejecting Rfam hits. There are likely Rfam hits where are
in a cluster, but do not overlap another sequence enough to be properly
rejected. For example:

|--------|           Rfam
       |-----------| SILVA

Here the Rfam would be rejected, but really we should only do it in
cases like:

|--------|   Rfam
  |--------| SILVA

It should require a high overlap but does not now.

Additionally, we have databases like GeneCards that use our annotations
to build things. Since they do not provide new sequences they should
probably (maybe?) not be counted when deciding if a sequence comes from
Rfam only. There may be further RNA type specific logic for this as
well.

2. Properly grouping sequences by RNA type is flawed. The normalizing
step does a good job with dealing with SSU vs 16S_rRNA and so forth but
does not deal with cases where overlapping sequences are SSU and rRNA.
We may want to group these things together or not. Consider a case like:

|-----------|  rRNA
 |-----------| SSU

Here we want both to be put together and treated as an SSU, probably. IT
will not happen yet. There are much more complex cases as well like:

|-----------------------------| rRNA
 |-------|                      SSU
                 |------------| LSU

Examples of this I have seen are that the large rRNA location is
actually the precursor of both. So there should be 3 groups, the rRNA,
the SSU and the LSU. This may not happen reliable currently. Also the
previous two cases can be combined and there is no way the correct
decisions are being made right now.

Also, there are times this may build two clusters, but when selecting
which things to highlight it needs to consider both clusters. For
example consider a cluster of 5.8S next to an LSU. We want to select a
5.8S  and LSU which are both 'good' and next to each other, no gaps
(unless that is common in that species). I'm not 100% sure if this
happens but if it does it would be from PDBe data, which is very
valuable and needs to handled correctly, so it is worth thinking about.

Finally there are problems caused by bad RNA types and QA limitations,
but the gene building step should help me find these cases.

In short, I think this redesign is much better and prevents some bugs
but I now need to get a corpus of actual examples to test against.