Finally fix issue with duplicate locations
This finally fixes the processing so that it seems to always assign a
location to a single cluster, fixing the issue where a location would be
in more than one cluster in some cases.

As a very nice side effect this has sped up the process quite a bit the
original version took hours on human and now this takes about 10 minutes
(on my laptop). Memory usage also appears reasonable, unlike some
previous attempts.

Interestingly, I more or less designed this the way I would in rust,
which helped fix the problem.