Create QA scan and genome mapping pipelines
I thought I was going to do the repeat analysis in these pipelines, only
to realize it probably isn't a good idea, it should happen in the QA
part of precompute as there is where we assign the QA status to things.
Since DSL v2 is likely the future of nextflow, I keep and commit the
code for future use anyway.