Outline representation of repeat regions
This shows more or less how I will represent and store information about
repetitive regions in genomes. The idea will be to fetch the soft masked
files from Ensembl and use that to extract the repetitive information.
From there we create tress to represent the ranges that are repetitive
and store them. This will be used in precompute as another QA check.