Outline parsing approach
The idea will be to use the incremental parsing by ijson to parse each
entry at a time and then build up the iterable of Entry objects. This
is just an outline of what should probably happen and will likely have
to be fixed up later.

Also fix location of some code.