Rework how repeats are found and represented
The previous two methods, a dict of IntervalTree and shelve stored
IntervalTrees, of building a RepeatTree had serious memory and time
issues. I can't use them as they would require too much memory (> 20GB)
or too much time (>1 day) to build the context. This method should be
much faster since it only relies upon files and some command line tools.

The idea is to use the query that Ensembl has provided to build a bed
like file for each assembly. This can then be indexed by tabix to
provide a fast lookup method for precompute. We use a series of these
files as a way to store and query the repeats instead of in memory data
structures.

This is likely slower than the IntervalTree in memory approach, but at
least it shouldn't eat up all the memory. Additionally, it doesn't take
long to build, a few minutes per species and can be parallelized easily.
This should fix the issues I'm having with repeats. The only issue I can
see is that if the file system is slow then precompute will be even
slower because each qa check involves file io. That said it would be
slow anyway so I don't think this is too much of a problem.

Currently untested but covers a better design.