Try to fix gtrnadb issue
For some reason the files we had for gtrnadb were incomplete. In order
to fix that I decided it would be a good idea to just fetch from the url
they provide. This adds some logic for going from the url they provide
to the actual files to download. We then use those urls to fetch all
files and then can parse the files. We should end up with all data
processed this way. This change is only for the V2 pipeline, but I
intend to move that pipeline for the next release.