Move repeats to own pipeline
It is complex enough that I want it in it's own pipeline. Also, I've
split the current logic into two parts, one to query ensembl one to
create the tabix files. The reason being that it is currently slow,
likely because I must limit the number of processes that hit the ensembl
database at once. However, I'm not sure and splitting things up this way
will help me time them. I may return to parsing fasta file for speedups.
This may be a slower step but it could be run in parallel more easily,
which would help things a lot.