Work to add ensembl rapid release
There are lots of genomes that E! puts in their rapid release section,
we can probably parse and import all this data by simply parsing the
embl files they generate. This alters the logic to also fetch the
species information for E! rapid release as well. This means the files
will be parsed and sequences imported. No genome information is fetch so
we will no do more genome mapping. Not sure if that is good or not
though.