A new way of fetching ENA wgs files
We should fetch wgs files differently because there are so many. In
general, we need to ensure there are no more than 4k files in a
directory (some nfs limit) and we want to work in parallel as much as
possible. Also we want to keep the total number of ENA jobs to something
reasonable, roughly in the 10,000s. This does that.

We find all wgs files we could use and then fetch the directory in
parallel. We then group the files into a total number of 4000 or less
and concatenate all files together, uncompressed, into the final files.
This should then give a reasonable number of total jobs, 20k (50
directories with 4000 files in each), with no directory that we may need
to work with having more than 4000 files in each. Also, each files will
hopefully be a reasonable size, not taking too long to process. Hard to
tell without a full fetch of the data though.

Also, note the other directories, tls, tsa, con-std, and such don't need
anything special and so we just fetch them all and process them as
needed.