First attempts at excluding metagenome rRNA
This adds some logic for trying to exclude metagenome rRNAs. The idea is
that if a sequence is an rRNA and comes from a metagenome it must have a
large hit to an rRNA model otherwise it must be excluded. Right now it
requires >90% coverage to the model, otherwise it is excluded.