Improve ENA DR line caching
Previously I cached all DR line information in memory. Because of this I
had to request workers with ~20GB of memory for parsing ENA data. This
is excessive, makes queueing slower and is generally weird. I added
sqlitedict which will store the data on disk in an sqlite database
instead. This will reduce the memory usage, but it does mean we have to
have a performant file system to work well. That would probably be true
anyway as we are doing IO but this would make the issue worse.

Also, update tests to use new API.