Rework ENA parsing to use a context
It was getting too complex with the ribovore results, DR line mapping
and TPA mappings. So I collected them all into a single context object.
This makes things a bit easier to follow. Also create a helpers.to_entry
as this is a nice pattern I use elsewhere.

Also, add more types to the TPA module.