Improve logic to get ensembl paths
We need to figure out what parts of the FTP dump to download. This
switches to having some python logic to build the paths. The idea is
that we need to do this in order to correctly get the GFF3 paths. We
don't want all gff3 files and only want the top level one. This
generates that path, along with the paths to latest ensembl data files.