Start adding GFF3 parsing to Ensembl data
We have to use GFF3 files for extracting Ensembl data now. Turns out
that the coordinates we currently use are relative to the sequence in
the file, which can differ from the genome as they may collapse and
remove parts. However, the GFF3 files have all the correct coordinates
are are pretty easy to parse so we can use them instead. To do that we
need to get the URL for the GFF3 files and download them along with the
embl files we currently parse.

This reorganizes lots of code along the way. I prefer to keep the
structure of the code matching the structure of databases, so putting
all the Ensembl databases together and renaming things to reflect the
fact that E!Genomes is now part of Ensembl proper is nice. Also, what is
generally called Ensembl is now often referred to as Ensembl vertebrates.

It also turns out the GFF3 files contains SO term annotations for all
sequences so we can use that as well. I think the entire parsing logic
can be moved to just using GFF3/fasta files, but I don't want to do that
deep of a rewrite yet.

Also, the E! GFF3 file may contain all information from GENCODE and
indicate which transcripts are from GENCODE. If so we can remove the
logic to fetch the GENCODE data and just use the standard GFF3 file.
Still have to check that though.