More merge of E!genomes and ensembl vert
Basically these two resources are so similar I can merge lots of the
code. And I should. I've moved some code around to reflect this. This
also adds some logic to fetch urls for data to get for E! genomes. This
is useful for fetching gff files for Ensembl parsing.