Try to split overly large ENA files
Some ENA files are so large that we struggle to import. We need to split
these large files apart. This add some logic to try and do that. The
exact cutoffs aren't selected yet though. This just outlines what needs
to happen.

I also modify the split-sequences script to work on ENA files as well.
Hopefully BioPython can do a loseless parse of these files. Otherwise
things will have to be a bit trickier.