Add some logic to track and write counts
We want to track how many things are excluded and why while parsing ENA
data. This adds some logic to do so. We also copy the data into a final
directory, which must be created ahead of time. This counting is likely
a temporary thing so the solution isn't totally worked out.