Fix e!vert parsing
This now uses the new GFF files to get the corrected coordinates and
GENCODE information. We want to use the GFF3 files as the source for
good genomic information and some other annotations. This parses the
GFF3 files for that information and uses. I could probably merge ensembl
and ensembl genomes data more, but for now this is ok.