Try to restructure SILVA parsing
SILVA parsing is super slow. I suspect this is because it has to
constantly request taxonomic information from ENA. We can fix this by
just indexing and using the data we fetch from the NCBI. This reworks
SILVA parsing to store such data and then access it as needed. It will
still try to fetch such data from the ENA service, but only after it
fails from the indexed data. Hopefully this helps speed it up.