Provide fallback data for nonchromosomal sequences
We parse non-chromosomal sequences and when this happens they do not
have entries in the GFF file. We still want to parse and use this data
so we will just fallback to the previous method. This may produce some
'bad' locations for these sequences though. I could also skip creating
any coordinates, but then they would mapped anyway which has it's own
issues.