Improve inference of parameters
The flags based upon which databases are active needed to be improved.
First, the name for the one that indicates if publications are to be
fetched was unclear. Secondly, neither one dealt with ensembl correctly.
Basically they didn't deal with nested data correctly. I didn't write
the proper search logic, just hardcoded ensembl for now. That should be
ok I hope.