Start building a context
I'm going to need a common context object. For some things we need to
lookup taxonomic information, for others SO term info. I will store this
in a context that gets used for most parsers. This will help me prevent
some hardcoded data and behaviors. For now it will stick to just
taxonomic information despite how it is named.