Improve precompute
This now makes sure that SO term logic will always produce some RNA
type. It has some more rules and will fallback to using ncRNA instead of
None if something is unclear. The logic isn't perfect yet, but is mostly
tolerable. It is also a clear improvement in the tested cases.

Additionally, this fixes other issues with precompute so it can actually
run now.