Try to split apart precompute queries
This splits the queries into parts and modifies the pipeline to use
them. We do each query separately and then index all data with kv. Then
we process each chunk separately. I think this will fix some of the
issues we are having with precompute as we won't run into as many
database issues from doing all those joins.