Rework precompute id lookup
The precomptue id lookup method is too slow and memory intensive. This
reworks it to be done outside the database which should limit the memory
pressure. The general idea is that we produce CSV's of the urs and
release id, we then use this to select all URS where the largest release
id is bigger than the last time it was computed, or if it was never
computed or is new. This is basically done by sorting CSV's and
joining them easily since they are sorted.