Expand urs to urs taxid for precompute
Precompute the logic works on the basis of a urs_taxid pair, that is wed
compute information for a specific sequence in a specific species.
However, we decide what sequences to compute based upon the URS, that is
we recompute all instances if any instance is changed. This isn't ideal
and can probably be fixed but it is how things work now. Also it is much
faster to compute things based upon the general URS instead of the
specific ones. This means that we need to expand the general URS into
the specific ones. Thus I added some scripts to do this. The script to
do this is not optimized at all, it could be much faster if I sorted
things and the like, but I haven't done that yet. I may need too
however.

Also had to fix some bugs and work around the fact that a process can't
be used more than once in a nextflow workflow context.