Try switching logic of precompute querying
So right now it is way too slow. I think this can be dealt with by not
trying to index all the accession information as that takes too long.
Instead we can query the accessions and use indexing for the rest of the
data. This should preform well, or at least not too awfully.

However, there is one issue with the querying by accessions, doing a join between
xref and rnc_accessions is way too slow and memory draining. Instead we
build a table that does that ahead of time. We can build it in parts so
there isn't too much strain on the DB and it will happen reasonably
fast. This is pretty complex so I had to build a new pipeline just for
that part however.