Rework precompute workflow
This reworks the entire workflow to use the new method. The new method
has a different approach.

First we begin by figuring out all urs_taxid pairs to process. This is
basically all urs that have been changed or are new, expanded to all
urs_taxid pairs for the urs. This is put in a table for later work, here
I will call it the todo table.

The todo table has one entry for each urs_taxid and has a sequential int
key we can use for processing. We build a file of chunks from this table,
where each chunk is reasonable number of urs_taxids to process
independently. We want several chunks otherwise it will take forever to
actually do precompute, so we need to get largish chunks. Normally
getting ~50-100 chunks total is reasonable.

Next we build up a table that is the result of an xref/accession join
for all entries in the todo table. We can't do the whole join at once,
it will never work. So instead we do it one partition at a time, this is
pretty fast right now (~2 hours) and actually finishes. We use this
table later in place of the accession table for querying.

While that is happening we pull out all metadata, ie xrefs, rfam hits,
etc, for all entries in the todo table. This is also fast (~1hour)
currently. We write out in sorted order so that we can then join the
files quickly. We build up what is basically the partial data for
precomptue this way. We then split the merged metadata file into chunks
using the chunk information above.

Then we query the built accession table for each chunk and merge it with
the metadata chunk and do precompute for that range. Once they are all
done we load the information and do a little cleanup.

That is the idea at least. These changes are totally tested yet, as I'm
doing this manually for now, but will be later.