Improve logic of building precompute tables
We need to do a better job at building the tables. First we need to
bring back the ability to select a method to find ids. Secondly we need
to ensure that the sorting is done in a better order. Specifically we
want to order things so that all URS and won't get broken up at the
edges of some chunk. This does that.