Move xref info around
We don't actually need to query against the xref table, which is slow.
We can use the generated accession table which basically denormalizes
the xref and accession table. By doing this we get rid of the slowest
and most expensive query.

Instead we create a basic object which contains the length and urs/taxid
information. This replaces xref for building the merged data.