Add some grouping/validation
This should help validation of the data. Basically the hope is that this
way we can ensure that all data which should be written is. It will also
make it is clearer when we join things. We will have one entry per
possible output so joining will just be iterating over all lines at
once.