Use pregrouping logic
The idea is that before we build up the final metadata file we can group
query results so there is one entry per line and one line per entry. By
doing this all files should have the same number of lines and not miss
any entries. This makes the logic of joining files basically trivial as
well as providing several places for better error checking.