Add grouping to search export
I am going to try out this trick for search export. I think it will help
reduce the amount of code and make things just look nicer in general.

To do this I had to move the grouper code from precompute into rnc-core
and then add the needed bits to search-export. Most of the changes are
pretty mechanical.