Ensure we use chromosome MT, not M
We refer to human chromosomes using ensembl names, This can differ from
the ones LncBook provides in that sometimes they give M for the
mitochromosome. This fixes that so we will always use MT instead. It
would be nice to have some db constraints that ensure these things
automatically, but not sure where to get all that data yet.