Try only build channels when needed
The issue with the previous setup was that it would try to build id
and query channels every time. This isn't ideal as there may be no
configured id file or query leading to failures. Instead what this does
is try to only build the file object for the channels when needed. I
think it will work.