Ensure only human and mouse are GENCODE
Seems that ensembl uses havana annotations both for gencode (which is
human and mouse only) and havana. This is a problem because organisms
other than human and mouse have these annotations. We should only treat
sequences with havana annotations from human and mouse as gencode. The
other sequence will just be treated as normal ensembl annotations.