Deal with xrefs in only the xref table
Currently, HGNC data does not get inserted into the HGNC partition,
likely an issue with the trigger function. While that also needs to be
fixed for now I will just add a special case where we query from only
the xref table to build up new entries. This has to be a different case
because we can't constrain the database column in accessions as we do
normally. It should still be fast though. I hope. We shall see.