Figure out precompute_urs_taxid range
We need this when grouping accessions. We query accessions by urs id so
we get all urs in a single chunk by the grouping is done by urs_taxid.
To deal with this easily I figure out what the correct start/stop of
urs_taxid should be for the given urs_id range. This adds some logic to
do that in python.