Begin changes to merge all pipelines
We will want to have a single entry point for much of the pipeline. This
tends to make it easier to run things as well as know what has happened.
To do this I need to make sure that each part runs when it should so I
am passing around flags to indicate when things can run. This makes
changes to many of the workflows to ensure they run in the right order.