Add more sane writers
The current setup for creating writers is needlessly complex. With just
a small amount of boiler plate we can make it a nice and clear system.
It covers the features from the old approach and is easier to read and
has less code. Untested but this is the setup I will go with from now on.