Add HGNC test
Basically this adds some test and test data for HGNC. I dumped all
current mappings and we can use that as the test data as it is based off
the old code. This will ensure I don't introduce new bugs. I've started
fixing the bugs I've found because of the tests but it mostly works now.